changes all time-variables for pomodoro from double to int reason: Duration class of Dart convert e.g. minutes [int] to correct duration

+++ lust/lib/widgets/pomodoroPage/pomodoroDescription.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 
 class pomodoroDescription extends StatelessWidget {
-  double periodTime;
-  double shortBreakTime;
-  double longBreakTime;
+  int periodTime;
+  int shortBreakTime;
+  int longBreakTime;
   int countPeriods;
 
-  pomodoroDescription(double periodTime, double shortBreakTime, double longBreakTime, int countPeriods){
+  pomodoroDescription(int periodTime, int shortBreakTime, int longBreakTime, int countPeriods){
     this.periodTime=periodTime;
     this.shortBreakTime=shortBreakTime;
     this.longBreakTime=longBreakTime;
